fix(builder-rsbuild): keep previewMainTemplate in core presets
